refactor(core): extract createCursorOverride
Pull the document-level cursor stylesheet injection out of init() into
its own module. The factory takes the three phase accessors that
determine the cursor (activated, copying, promptMode), owns the lazy
`<style data-react-grab-cursor>` element and its CSP nonce + third-
party-hide setup, and registers its own createEffect.

`cursorOverride.clear()` replaces the dispose-path `setCursorOverride(null)`
call.

Drops 2 now-unused imports (detectCspNonce, hideFromThirdParties).

index.tsx: 3218 -> 3185 lines (-33). No behavior change.
